+++ azure-storage-queue/azure/storage/queue/_constants.py
@@ -8,4 +8,4 @@ __author__ = 'Microsoft Corp. <ptvshelp@microsoft.com>'
 __version__ = '1.0.0'
 
 # x-ms-version for storage service.
-X_MS_VERSION = '2017-04-17'
+X_MS_VERSION = '2017-07-29'