+++ tests/blob/test_common_blob.py
@@ -12,6 +12,7 @@ from datetime import datetime, timedelta
 from azure.common import (
     AzureHttpError,
     AzureMissingResourceHttpError,
+    AzureException,
 )
 from azure.storage.common import (
     AccessPolicy,
@@ -443,6 +444,34 @@ class StorageCommonBlobTest(StorageTestCase):
         self.assertEqual(blob.properties.blob_type, self.bs.blob_type)
         self.assertEqual(blob.properties.content_length, len(self.byte_data))
         self.assertEqual(blob.properties.lease.status, 'unlocked')
+
+    # This test is to validate that the ErrorCode is retrieved from the header during a
+    # HEAD request.
+    @record
+    def test_get_blob_properties_fail(self):
+        # Arrange
+        blob_name = self._create_block_blob()
+
+        # Act
+        with self.assertRaises(AzureException) as e:
+            self.bs.get_blob_properties(self.container_name, blob_name, 1) # Invalid snapshot value of 1
+
+        # Assert
+        self.assertIn('ErrorCode: InvalidQueryParameterValue', str(e.exception))
+
+    # This test is to validate that the ErrorCode is retrieved from the header during a
+    # GET request. This is preferred to relying on the ErrorCode in the body.
+    @ record
+    def test_get_blob_metadata_fail(self):
+        # Arrange
+        blob_name = self._create_block_blob()
+
+        # Act
+        with self.assertRaises(AzureException) as e:
+            self.bs.get_blob_metadata(self.container_name, blob_name, 1) # Invalid snapshot value of 1
+
+        # Assert
+        self.assertIn('ErrorCode: InvalidQueryParameterValue', str(e.exception))
 
     @record
     def test_get_blob_server_encryption(self):