+++ azure-storage-common/ChangeLog.md
@@ -1,6 +1,8 @@
 # Change Log azure-storage-common
 
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
+## Version XX.XX.XX:
+- Error message now contains the ErrorCode from the x-ms-error-code header value
 
 ## Version 1.0.0: